WIP add ObjectValue
