+++ packages/data-mate/src/core/index.ts
@@ -2,6 +2,7 @@ export * from './config';
 export * from './DateValue';
 export * from './interfaces';
 export * from './IPValue';
+export * from './ObjectValue';
 export * from './ReadableData';
 export * from './utils';
 export * from './WritableData';
