+++ packages/data-mate/src/core/utils.ts
@@ -117,6 +117,21 @@ function _createObjectHashCode() {
     return getHashCodeFrom(Object.entries(this), false);
 }
 
+export function freezeObject<T extends Record<string, any>>(
+    input: T
+): T extends Readonly<infer U> ? Readonly<U> : Readonly<T> {
+    if (Object.isFrozen(input)) return input as any;
+    return Object.freeze({ ...input }) as any;
+}
+
+type ArrLike = (any[])|(readonly any[]);
+export function freezeArray<T extends ArrLike>(
+    input: T
+): T extends Readonly<infer U> ? Readonly<U> : Readonly<T> {
+    if (Object.isFrozen(input)) return input as any;
+    return Object.freeze(input.slice()) as any;
+}
+
 export function getObjectDataTypeConfig(
     config: DataTypeFields|ReadonlyDataTypeFields, baseField: string
 ): DataTypeFields {
